*: move pkg/colorutil to pkg/utils
