+++ pkg/logger/debug.go
@@ -20,8 +20,8 @@ import (
 	"path/filepath"
 	"time"
 
-	"github.com/pingcap/tiup/pkg/colorutil"
 	"github.com/pingcap/tiup/pkg/localdata"
+	"github.com/pingcap/tiup/pkg/utils"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 )
@@ -52,9 +52,9 @@ func OutputDebugLog(prefix string) {
 
 	err := os.WriteFile(filePath, debugBuffer.Bytes(), 0644)
 	if err != nil {
-		_, _ = colorutil.ColorWarningMsg.Fprint(os.Stderr, "\nWarn: Failed to write error debug log.\n")
+		_, _ = utils.ColorWarningMsg.Fprint(os.Stderr, "\nWarn: Failed to write error debug log.\n")
 	} else {
-		_, _ = fmt.Fprintf(os.Stderr, "\nVerbose debug logs has been written to %s.\n", colorutil.ColorKeyword.Sprint(filePath))
+		_, _ = fmt.Fprintf(os.Stderr, "\nVerbose debug logs has been written to %s.\n", utils.ColorKeyword.Sprint(filePath))
 	}
 	debugBuffer.Reset()
 }
